backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.10.0
anthropic==0.18.1
google-generativeai==0.3.2
python-dotenv==1.0.0
httpx==0.26.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10
